+++ pcap4j-sample/src/main/java/org/pcap4j/sample/DumpLoop.java
@@ -2,6 +2,7 @@ package org.pcap4j.sample;
 
 import java.io.IOException;
 import org.pcap4j.core.BpfProgram.BpfCompileMode;
+import org.pcap4j.core.NotOpenException;
 import org.pcap4j.core.PcapDumper;
 import org.pcap4j.core.PcapHandle;
 import org.pcap4j.core.PcapNativeException;
@@ -9,11 +10,6 @@ import org.pcap4j.core.PcapNetworkInterface;
 import org.pcap4j.core.PcapNetworkInterface.PromiscuousMode;
 import org.pcap4j.util.NifSelector;
 
-/**
- *
- * @author Kaito
- *
- */
 public class DumpLoop {
 
   private static final String COUNT_KEY
@@ -36,12 +32,7 @@ public class DumpLoop {
   private static final String PCAP_FILE
     = System.getProperty(PCAP_FILE_KEY, "DumpLoop.pcap");
 
-  /**
-   *
-   * @param args
-   * @throws PcapNativeException
-   */
-  public static void main(String[] args) throws PcapNativeException {
+  public static void main(String[] args) throws PcapNativeException, NotOpenException {
     String filter = args.length != 0 ? args[0] : "";
 
     System.out.println(COUNT_KEY + ": " + COUNT);